ESLint rules. Turning off two rules for now since they're widely used and affect functionality
