+++ examples/src/demo-transaction.js
@@ -1,16 +1,16 @@
 export default {
   actions: [{
-    account: "eosio.token",
-    name: "transfer",
+    account: 'eosio.token',
+    name: 'transfer',
     authorization: [{
-      actor: "", // use account that was logged in
-      permission: "active",
+      actor: '', // use account that was logged in
+      permission: 'active',
     }],
     data: {
-      from: "", // use account that was logged in
-      to: "cryptkeeper",
-      quantity: "1.0000 EOS",
-      memo: "UAL works!",
+      from: '', // use account that was logged in
+      to: 'cryptkeeper',
+      quantity: '1.0000 EOS',
+      memo: 'UAL works!',
     },
   }],
 }
