--- examples/tslint.json
@@ -1,16 +1,0 @@
-{
-  "defaultSeverity": "error",
-  "extends": [
-    "@blockone/tslint-config-blockone/tslint-config"
-  ],
-  "jsRules": {},
-  "rules": {
-  },
-  "rulesDirectory": [],
-  "linterOptions": {
-    "exclude": [
-      "src/**/templates/",
-      "**/*.js"
-    ]
-  }
-}
